Gedder Engine now works as intended. See details of commit message.
We now check what the "optimal wake up time" is, meaning
the time that we HAVE to wake up given what we just detected
from the Gedder Engine, as well as the prep time and what we
previously knew about the arrival time.

We then see whether the current time has PASSED that optimal
wake up time. If it has, we obviously wake up.

Otherwise, just check back again given some frequency that
we determine based off of the time until alarm.