Remove redundant variable "upperBound". See details.
Due to a misunderstanding on how the gedder algorithm should work,
we ended up implementing a time separate to alarm time, and called it
upper bound time. But the algorithm makes no use of this and is
essentially unnecessary since the alarm time is not actively
changing.

So, we remove all appearances of it throughout the application.